Service.

+++ src/main/java/hu/elte/alkFejlBead/alkFejlBead/repositories/DijcsomagRepository.java
@@ -1,9 +1,9 @@
 package hu.elte.alkFejlBead.alkFejlBead.repositories;
 
 
-import hu.elte.alkFejlBead.alkFejlBead.entities.Dijcsomag;
+import hu.elte.alkFejlBead.alkFejlBead.Model.Dijcsomag;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DijcsomagRepository extends CrudRepository<Dijcsomag, Integer> {
+public interface DijcsomagRepository extends CrudRepository<Dijcsomag, Long> {
 
 }
