+++ src/main/java/hu/elte/alkFejlBead/alkFejlBead/repositories/DolgozoRepository.java
@@ -1,7 +1,8 @@
 package hu.elte.alkFejlBead.alkFejlBead.repositories;
 
-import hu.elte.alkFejlBead.alkFejlBead.entities.Dolgozo;
+
+import hu.elte.alkFejlBead.alkFejlBead.Model.Dolgozo;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DolgozoRepository extends CrudRepository<Dolgozo, Integer> {
+public interface DolgozoRepository extends CrudRepository<Dolgozo, Long> {
 }
