+++ src/main/java/hu/elte/alkFejlBead/alkFejlBead/repositories/SzerzodesRepository.java
@@ -1,7 +1,8 @@
 package hu.elte.alkFejlBead.alkFejlBead.repositories;
 
-import hu.elte.alkFejlBead.alkFejlBead.entities.Szerzodes;
+import hu.elte.alkFejlBead.alkFejlBead.Model.Szerzodes;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SzerzodesRepository extends CrudRepository<Szerzodes, Integer> {
+
+public interface SzerzodesRepository extends CrudRepository<Szerzodes, Long> {
 }
